+++ backend/src/Modules/Plans/Planner.Plans.Application/Plans/DeletePlan/DeletePlanCommand.cs
@@ -3,4 +3,4 @@ using Planner.Common.Domain.Core.Primitives.Result;
 
 namespace Planner.Plans.Application.Plans.DeletePlan;
 
-public sealed record DeletePlanCommand(int UserId, Guid PlanId) : ICommand<Result>;
+public sealed record DeletePlanCommand(long UserId, Guid PlanId) : ICommand<Result>;
