+++ backend/src/Modules/Plans/Planner.Plans.Application/Plans/GetUserPlans/GetUserPlansQuery.cs
@@ -2,4 +2,4 @@ using Planner.Common.Application.Messaging;
 
 namespace Planner.Plans.Application.Plans.GetUserPlans;
 
-public sealed record GetUserPlansQuery(int UserId) : IQuery<IReadOnlyCollection<PlanDto>>;
+public sealed record GetUserPlansQuery(long UserId) : IQuery<IReadOnlyCollection<PlanDto>>;
