+++ backend/src/Modules/Plans/Planner.Plans.Domain/Plans/Repositories/IPlanRepository.cs
@@ -6,7 +6,7 @@ public interface IPlanRepository
 {
     Task<Plan?> TryGetByIdAsync(Guid planId, CancellationToken cancellationToken);
 
-    Task<Plan?> TryGetByAsync(Guid planId, int userId, CancellationToken cancellationToken);
+    Task<Plan?> TryGetByAsync(Guid planId, long userId, CancellationToken cancellationToken);
 
     Task InsertAsync(Plan entity, CancellationToken cancellationToken);
